refactor: extract billing cleanup logic to reduce overall code complexity
Move checkout/invoice cleanup helpers (getStripeObjectId, getCheckoutRefundTarget,
cleanupBlockedCheckoutSession, refundDeletedAccountInvoice, and related Stripe
payment resolution functions) into billing-cleanup.js to reduce billing.js
overall code complexity below the CodeScene threshold.
